fix exception when undefined value
